Sun, Dec 18, 2022, 8:31 AM +00:00

+++ pset2/readability/readability.c
@@ -7,7 +7,7 @@
 int count_letters(string text);
 int count_words(string text);
 int count_sentences(string text);
-int cal_index(int letters, int words, int sentences);
+double cal_index(int letters, int words, int sentences);
 
 int main(void)
 {
@@ -18,7 +18,7 @@ int main(void)
     letters = count_letters(text);
     words = count_words(text);
     sentences = count_sentences(text);
-    index = cal_index(letters, words, sentences);
+    index = (int)cal_index(letters, words, sentences);
 
     if (index < 1)
     {
@@ -81,10 +81,10 @@ int count_sentences(string text)
     return count;
 }
 
-int cal_index(int letters, int words, int sentences)
+double cal_index(int letters, int words, int sentences)
 {
     double l = letters / words * 100;
     double s = sentences/ words * 100 ;
 
-    return (int)round(0.0588 * l - 0.296 * s - 15.8);
+    return round(0.0588 * l - 0.296 * s - 15.8);
 }
